For the "In-Lecture Screen", we can rotate the phone to landscape mode and have a larger size for the slides and still make confusionMarks and notes via the buttons as before. Switching between portrait (MainActivity) and landscape (LandscapeInLectureActivity) persists whatever lecture and slide number the user is currently using via "LectureId" and "SlideNumber" in the Intents that start those Activities.
Switching between landscape and portrait mode for the in-lecture screen removes the previous activity from the top of the stack via .finish() (so hitting the back button at any time goes back to the lecture-list activity)

Note: In order to have the rotation work in my emulator, I had to enable "Auto-rotate" in the emulator. For a Pixel XL, you can do this by swiping down the notification bar and selecting the auto-rotate icon.