Ability to sort questions and timestamps via a spinner (spinner purposely not included in activity_main.xml to prevent merge conflicts...but TODO will later have to be hooked up to some spinner with the sort_list_menu_choices string-array included in this commit's strings.xml)
QuestionListFragment and TimestampsListFragment maintain additional state: the criteria by which they are currently sorting their elements. Hacky but it works

Refactored how refreshing works in QuestionsListFragment and TimestampsListFragment

All buttons in the ButtonFragment will switch the list fragment and set the sort spinner to "Date Created" when their respective entities are inserted.

Fixed bug where when you opened the MainActivity for the first time, the timestamps list fragment was created but the "Questions" option in the bottom navigation bar was selected